+++ app/javascript/dashboard/helper/commons.js
@@ -61,3 +61,10 @@ export const createPendingMessage = data => {
 
   return pendingMessage;
 };
+
+export const convertToSlug = text => {
+  return text
+    .toLowerCase()
+    .replace(/[^\w ]+/g, '')
+    .replace(/ +/g, '_');
+};
