+++ app/javascript/dashboard/routes/dashboard/settings/attributes/AddAttribute.vue
@@ -85,7 +85,7 @@
 <script>
 import { required, minLength } from 'vuelidate/lib/validators';
 import { mapGetters } from 'vuex';
-import { convertToSlug } from '../../../../../widget/helpers/convertToSlugHelper';
+import { convertToSlug } from 'dashboard/helper/commons.js';
 import { ATTRIBUTE_MODELS, ATTRIBUTE_TYPES } from './constants';
 import alertMixin from 'shared/mixins/alertMixin';
