--- app/javascript/widget/helpers/convertToSlugHelper.js
@@ -1,6 +1,0 @@
-export const convertToSlug = text => {
-  return text
-    .toLowerCase()
-    .replace(/[^\w ]+/g, '')
-    .replace(/ +/g, '_');
-};
