--- app/javascript/widget/helpers/specs/convertToSlugHelper.spec.js
@@ -1,8 +1,0 @@
-import { convertToSlug } from '../convertToSlugHelper.js';
-describe('#Attribute Helper', () => {
-  describe('convertToSlug', () => {
-    it('should convert to slug', () => {
-      expect(convertToSlug('Test@%^&*(){}>.!@`~_ ing')).toBe('test__ing');
-    });
-  });
-});
